+++ CHANGELOG.md
@@ -7,9 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
-- `QasNestedFields` new prop returnRow.
-- `QasNestedFields` returning changed line index.
-- `QasNestedFields` enumerating label.
+- `QasNestedFields` changes to get row index.
 - Mask money in `filters.js`
 - Mask square area in `filters.js`
 - Mask decimal in `filters.js`