+++ CHANGELOG.md
@@ -12,9 +12,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New mixin mapMarkerMixin
 - Create class `.opaque`
 - Plugin Autoprefixer
+- `QasNestedFields` changes to get row index.
+- Mask money in `filters.js`
+- Mask square area in `filters.js`
+- Mask decimal in `filters.js`
 
 ### Fixed
 
+### Fixed
 - `Qastip` computed and close icon ajusted.
 
 ## 2.6.1 - 2021-07-01